% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CleanExportedMOMOData.R
\name{CleanExportedMOMOData}
\alias{CleanExportedMOMOData}
\title{CleanExportedMOMOData}
\usage{
CleanExportedMOMOData(data, s,
  folder_results = fhi::DashboardFolder("results"))
}
\arguments{
\item{data}{a}

\item{s}{a}

\item{folder_results}{a}
}
\description{
CleanExportedMOMOData
}
