% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{CreateLatestDoneFile}
\alias{CreateLatestDoneFile}
\title{test}
\usage{
CreateLatestDoneFile(data_clean = fhi::DashboardFolder("data_clean"), f)
}
\arguments{
\item{data_clean}{a}

\item{f}{a}
}
\description{
test
}
