% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EmailNotifiation.R
\name{EmailInternal}
\alias{EmailInternal}
\title{blah}
\usage{
EmailInternal(folderResultsYearWeek, isTest = TRUE)
}
\arguments{
\item{folderResultsYearWeek}{a}

\item{isTest}{a}
}
\description{
blah
}
