% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GenerateStack.R
\name{GenerateStack}
\alias{GenerateStack}
\title{GenerateStack}
\usage{
GenerateStack(folder_clean = DashboardFolder("data_clean"),
  folder_results = DashboardFolder("results"), f, dateDataMinusOneWeek,
  dateData)
}
\arguments{
\item{folder_clean}{a}

\item{folder_results}{a}

\item{f}{a}

\item{dateDataMinusOneWeek}{a}

\item{dateData}{a}
}
\description{
GenerateStack
}
