% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetDataInfo}
\alias{GetDataInfo}
\title{GetDataInfo}
\usage{
GetDataInfo(folder_raw = fhi::DashboardFolder("data_raw"),
  folder_clean = fhi::DashboardFolder("data_clean"))
}
\arguments{
\item{folder_raw}{a}

\item{folder_clean}{a}
}
\description{
GetDataInfo
}
