% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGraphs.R
\name{RunGraphsDeaths}
\alias{RunGraphsDeaths}
\title{Running graphs}
\usage{
RunGraphsDeaths(runName = "Norway", data,
  folder = fhi::DashboardFolder("results", file.path(RAWmisc::YearWeek(),
  "Graphs")), yearWeek = RAWmisc::YearWeek(), dateData = Sys.time(),
  dateReliable = RAWmisc::YearWeek())
}
\arguments{
\item{runName}{Name}

\item{data}{data}

\item{folder}{a}

\item{yearWeek}{a}

\item{dateData}{a}

\item{dateReliable}{a}
}
\description{
Running graphs
}
