% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RunGraphs.R
\name{RunGraphsStatistics}
\alias{RunGraphsStatistics}
\title{Running graphs}
\usage{
RunGraphsStatistics(runName = "Norway", allPlotData,
  folder = fhi::DashboardFolder("results", file.path(RAWmisc::YearWeek(),
  "Graphs")), yearWeek = RAWmisc::YearWeek(), dateData = Sys.time())
}
\arguments{
\item{runName}{Name}

\item{allPlotData}{allPlotData}

\item{folder}{a}

\item{yearWeek}{a}

\item{dateData}{a}
}
\description{
Running graphs
}
