% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatusTiles.R
\name{RunStatusTiles}
\alias{RunStatusTiles}
\title{blah}
\usage{
RunStatusTiles(allResults, folder, yearWeek, dateData)
}
\arguments{
\item{allResults}{a}

\item{folder}{a}

\item{yearWeek}{a}

\item{dateData}{a}
}
\description{
blah
}
