% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_temp_graphs.R
\name{RunTemporaryGraphs}
\alias{RunTemporaryGraphs}
\title{Run Temporary Graphs}
\usage{
RunTemporaryGraphs(runName = "Norway", masterData,
  folder = fhi::DashboardFolder("results", file.path(RAWmisc::YearWeek(),
  "Graphs")), yearWeek = RAWmisc::YearWeek(), dateData = Sys.time())
}
\arguments{
\item{runName}{Name}

\item{masterData}{masterData}

\item{folder}{a}

\item{yearWeek}{a}

\item{dateData}{a}
}
\description{
Running graphs
}
