% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SavingRawData.R
\name{SavingRawData}
\alias{SavingRawData}
\title{SavingRawData}
\usage{
SavingRawData(folder_results = fhi::DashboardFolder("results"),
  dateDataMinusOneWeek, masterData)
}
\arguments{
\item{folder_results}{a}

\item{dateDataMinusOneWeek}{a}

\item{masterData}{a}
}
\description{
SavingRawData
}
