% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SetupFolders.R
\name{SetupFolders}
\alias{SetupFolders}
\title{SetupFolders}
\usage{
SetupFolders(folder_results = DashboardFolder("results"),
  dateDataMinusOneWeek)
}
\arguments{
\item{folder_results}{a}

\item{dateDataMinusOneWeek}{a}
}
\description{
SetupFolders
}
