% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ZipResults.R
\name{ZipResults}
\alias{ZipResults}
\title{blah}
\usage{
ZipResults(folderResults = fhi::DashboardFolder("results"),
  folderResultsYearWeek = fhi::DashboardFolder("results",
  RAWmisc::YearWeek()), folderResultsZip = fhi::DashboardFolder("results",
  paste0("archive_", RAWmisc::YearWeek(), ".zip")),
  folderDataAppZip = fhi::DashboardFolder("data_app", paste0("archive_",
  RAWmisc::YearWeek(), ".zip")))
}
\arguments{
\item{folderResults}{a}

\item{folderResultsYearWeek}{a}

\item{folderResultsZip}{a}

\item{folderDataAppZip}{a}
}
\description{
blah
}
