% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{GetData}
\alias{GetData}
\title{GetData}
\usage{
GetData(folder_raw = fhi::DashboardFolder("data_raw"), fClean, f,
  forceRun = FALSE)
}
\arguments{
\item{folder_raw}{a}

\item{fClean}{a}

\item{f}{a}

\item{forceRun}{a}
}
\description{
GetData
}
