% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_temp_graphs.R
\name{RunTemporaryGraphs}
\alias{RunTemporaryGraphs}
\title{Run Temporary Graphs}
\usage{
RunTemporaryGraphs(runName = "Norway",
  folder = fhi::DashboardFolder("results", file.path(RAWmisc::YearWeek(),
  "Graphs")), yearWeek = RAWmisc::YearWeek())
}
\arguments{
\item{runName}{Name}

\item{folder}{a}

\item{yearWeek}{a}
}
\description{
Running graphs
}
